rename the api module to api_info
