+++ pkg/api/api_info/models.go
@@ -1,4 +1,4 @@
-package api
+package api_info
 
 import (
 	"github.com/sitehostnz/gosh/pkg/api"
